Move body of Process.commonWaitFor to generified commonWaitForCondition
